reduce replacement characters down from all unicode confusables to visually similar entities from ISO-8859-1
As we only need a single character to be replaced the break any hashing this will likely be more effective at cirumventing potential input filtering. The majority of unicode confusables will not be present in 'valid' names and as such can easilly be filtered. Limiting to the ISO-8859-1 character set will provide greater overlap with more 'common' name characters and greater overall entropy.